requirements.txt
pandas==2.1.4
scikit-learn==1.5
joblib==1.4.2
google-cloud-storage==2.16.0
numpy==1.26.4
